SOCINT-266 Use /survey endpoint for available fulfilments Removed remaining non-global functions from utils>View
